refactor(agent-platform): extract fcclient into shared fcvm package
Move the Firecracker REST client from fc-agentd/internal/fcclient to a new
shared package fcvm/fcclient so a sibling daemon (semgrep-scand) can reuse the
Firecracker primitives. Go internal/ visibility forbids a sibling from
importing fc-agentd/internal/..., so the client now lives in fcvm/ with public
Bazel visibility. Zero behavior change: the package, source, and tests are
identical; only the import path and BUILD label move.
