+++ projects/agent_platform/fcvm/fcclient/BUILD
@@ -3,8 +3,8 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "fcclient",
     srcs = ["client.go"],
-    importpath = "github.com/jomcgi/homelab/projects/agent_platform/fc-agentd/internal/fcclient",
-    visibility = ["//projects/agent_platform/fc-agentd:__subpackages__"],
+    importpath = "github.com/jomcgi/homelab/projects/agent_platform/fcvm/fcclient",
+    visibility = ["//visibility:public"],
 )
 
 go_test(
